+++ benchmark/init.js
@@ -1,0 +1,49 @@
+var Benchmark = require('benchmark'),
+    Hawkejs = require('../index.js'),
+    hawkejs = new Hawkejs(),
+    libpath = require('path');
+
+hawkejs.addViewDirectory(libpath.resolve(__dirname, '..', 'test', 'templates'));
+
+function createTestVariables() {
+	var result = {
+		a: 'a',
+		b: 'b',
+		c: 'c',
+		my: {
+			ref: {
+				to: {
+					alpha: 'alpha'
+				}
+			}
+		},
+		stuff: 'stuff',
+		activity: {
+			Teachers: [
+				{title: 1}
+			]
+		},
+		teacher: {
+			Links: [
+				{title: 'First link title', _href: 'href'},
+				{title: 'Second link title'}
+			]
+		},
+		empty_obj: {},
+		test: {
+			name: 'testname',
+			one : 1,
+			three: 3
+		}
+	};
+
+	return result;
+};
+
+module.exports = {
+	Benchmark : Benchmark,
+	Suite     : Benchmark.Suite,
+	Hawkejs   : Hawkejs,
+	hawkejs   : hawkejs,
+	createTestVariables
+};