+++ benchmark/start.js
@@ -1,0 +1,7 @@
+var bench = require('./init.js');
+
+console.log('Hawkejs');
+require('./00-hawkejs');
+
+console.log('\nRenderer');
+require('./05-renderer');